New settings screen and ability to add new role
