+++ src/caretogether-pwa/src/Settings/SettingsScreen.tsx
@@ -1,0 +1,20 @@
+import { ProgressBackdrop } from '../Shell/ProgressBackdrop';
+import useScreenTitle from '../Shell/ShellScreenTitle';
+import { useDataLoaded } from '../Model/Data';
+import { Box } from '@mui/system';
+import { RolesSection } from './Roles';
+
+export function SettingsScreen() {
+  useScreenTitle('Settings');
+
+  const dataLoaded = useDataLoaded();
+  return !dataLoaded ? (
+    <ProgressBackdrop>
+      <p>Loading settings...</p>
+    </ProgressBackdrop>
+  ) : (
+    <Box sx={{ paddingTop: 2 }}>
+      <RolesSection />
+    </Box>
+  );
+}
